feat: add userInfo inquiry and modify(#10)
