+++ backend-spring/build.gradle
@@ -47,7 +47,8 @@ dependencies {
 	implementation 'com.fasterxml.jackson.core:jackson-databind'
 
 	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.0'
-	// [추가] @ConfigurationProperties (KakaoProperties 등) 자동 완성을 위한 의존성
+
+	// @ConfigurationProperties (KakaoProperties 등) 자동 완성을 위한 의존성
 	annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
 }
