+++ backend-spring/src/main/java/com/mingleup/backend/domain/party/domain/PartyStatus.java
@@ -1,0 +1,8 @@
+package com.mingleup.backend.domain.party.domain;
+
+public enum PartyStatus {
+    RECRUITING, // 모집중
+    CLOSED,     // 모집 마감
+    SCHEDULED,  // 파티 예정중
+    COMPLETED   // 파티 종료
+}
