+++ backend-spring/src/main/java/com/mingleup/backend/domain/party/domain/RecruitmentMethod.java
@@ -1,0 +1,6 @@
+package com.mingleup.backend.domain.party.domain;
+
+public enum RecruitmentMethod {
+    APPROVAL, // 승인제
+    FCFS      // 선착순
+}
