+++ backend-spring/src/main/java/com/mingleup/backend/domain/user/domain/Gender.java
@@ -1,0 +1,5 @@
+package com.mingleup.backend.domain.user.domain;
+
+public enum Gender {
+    MALE, FEMALE, OTHER
+}
