+++ backend-spring/src/main/java/com/mingleup/backend/domain/user/domain/Role.java
@@ -1,0 +1,6 @@
+package com.mingleup.backend.domain.user.domain;
+
+public enum Role {
+    PARTICIPANT, // 참가자
+    HOST         // 호스트
+}
