--- backend-spring/src/main/java/com/mingleup/backend/domian/application/domain/ApplicationStatus.java
@@ -1,8 +1,0 @@
-package com.mingleup.backend.domian.application.domain;
-
-public enum ApplicationStatus {
-    PENDING,  // 승인 대기
-    APPROVED, // 승인
-    REJECTED, // 거절
-    ATTENDED  // 참석 완료
-}
