--- backend-spring/src/main/java/com/mingleup/backend/domian/party/domain/HostQuestion.java
@@ -1,34 +1,0 @@
-package com.mingleup.backend.domian.party.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "host_question")
-public class HostQuestion {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "question_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "party_id", nullable = false)
-    private Party party;
-
-    @Column(name = "question_text", nullable = false, columnDefinition = "TEXT")
-    private String questionText;
-
-    @Builder
-    public HostQuestion(Party party, String questionText) {
-        this.party = party;
-        this.questionText = questionText;
-        // 양방향 연관관계 편의 메서드 (선택)
-        // party.getHostQuestions().add(this);
-    }
-}
