--- backend-spring/src/main/java/com/mingleup/backend/domian/party/domain/RecruitmentMethod.java
@@ -1,6 +1,0 @@
-package com.mingleup.backend.domian.party.domain;
-
-public enum RecruitmentMethod {
-    APPROVAL, // 승인제
-    FCFS      // 선착순
-}
