--- backend-spring/src/main/java/com/mingleup/backend/domian/review/domain/ReviewType.java
@@ -1,8 +1,0 @@
-package com.mingleup.backend.domian.review.domain;
-
-public enum ReviewType {
-    PARTY,       // 파티 자체 후기
-    HOST,        // 호스트 후기
-    PARTICIPANT, // 참가자 후기
-    AI_GROUP     // AI 그룹 후기
-}
