--- backend-spring/src/main/java/com/mingleup/backend/domian/user/domain/Gender.java
@@ -1,5 +1,0 @@
-package com.mingleup.backend.domian.user.domain;
-
-public enum Gender {
-    MALE, FEMALE, OTHER
-}
