--- backend-spring/src/main/java/com/mingleup/backend/domian/user/domain/Role.java
@@ -1,6 +1,0 @@
-package com.mingleup.backend.domian.user.domain;
-
-public enum Role {
-    PARTICIPANT, // 참가자
-    HOST         // 호스트
-}
