+++ backend-spring/src/main/java/com/mingleup/backend/global/auth/dto/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.mingleup.backend.global.auth.dto; // 패키지 경로 수정
+package com.mingleup.backend.global.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -7,7 +7,7 @@ import lombok.ToString;
 
 /**
  * 카카오 사용자 정보 응답 DTO
- * [수정] 카카오 콘솔의 '필수 동의' 항목('name' 등)에 맞춘 버전
+ * [수정] 'profile_image' (선택 동의) 항목 추가
  */
 @Getter
 @ToString
@@ -17,10 +17,24 @@ public class KakaoUserInfoResponse {
     @JsonProperty("id")
     private String kakaoId;
 
-    // 'properties' (nickname, profile_image)는 "사용 안 함"이므로 DTO에서 제외
+    // [추가] 'profile_image'를 받기 위한 'properties'
+    @JsonProperty("properties")
+    private KakaoProperties properties;
 
     @JsonProperty("kakao_account")
     private KakaoAccount kakaoAccount;
+
+    // [추가] 'properties' 내부 클래스
+    @Getter
+    @ToString
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    public static class KakaoProperties {
+        @JsonProperty("profile_image")
+        private String profileImage;
+
+        @JsonProperty("thumbnail_image")
+        private String thumbnailImage;
+    }
 
     @Getter
     @ToString
@@ -31,7 +45,7 @@ public class KakaoUserInfoResponse {
         private String name; // "필수 동의" 항목 (이름)
 
         @JsonProperty("email")
-        private String email; // "선택 동의" 또는 "사용 안 함" (null일 수 있음)
+        private String email; // (null일 수 있음)
 
         @JsonProperty("gender")
         private String gender; // "필수 동의"
@@ -43,6 +57,22 @@ public class KakaoUserInfoResponse {
         private String birthday; // "필수 동의" (MMDD)
 
         @JsonProperty("birthyear")
-        private String birthyear; // "필S_ 동의" (YYYY)
+        private String birthyear; // "필수 동의" (YYYY)
+    }
+
+    // == 편의 메서드 == //
+
+    // [추가] 프로필 이미지를 가져오기 위한 편의 메서드
+    public String getProfileImageUrl() {
+        if (this.properties != null) {
+            if (this.properties.getProfileImage() != null) {
+                return this.properties.getProfileImage();
+            }
+            // 프로필 이미지가 없으면 썸네일 이미지 사용
+            if (this.properties.getThumbnailImage() != null) {
+                return this.properties.getThumbnailImage();
+            }
+        }
+        return null; // 둘 다 없으면 null 반환
     }
 }
